Rename `components/mdx/` to `components/theme/`
